Refactor Task and related classes into package
