+++ hamlet.txt
@@ -1,4 +1,4 @@
-T,1,work
-D,0,hw,2025-10-10
+T,0,work
+D,1,hw,2025-10-10
 E,0,part,2020-10-10,2020-10-10
 E,0,part,2025-10-10,2025-10-10
