+++ src/main/java/Hamlet.java
@@ -1,6 +1,7 @@
 package main.java;
 
 import main.java.exception.HamletException;
+import main.java.task.TaskList;
 
 import java.util.Scanner;
 import java.io.IOException;
