+++ src/main/java/Ui.java
@@ -1,5 +1,7 @@
 package main.java;
 
+import main.java.task.Task;
+
 import java.util.ArrayList;
 
 public class Ui {
