+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.task;
 
 public class Task {
     protected String description;
@@ -7,6 +7,10 @@ public class Task {
     public Task(String description) {
         this.description = description;
         this.isDone = false;
+    }
+
+    public String getDescription() {
+        return description;
     }
 
     public String getStatusIcon() {
